+++ tests/Instances/Control/Exception.hs
@@ -104,6 +104,12 @@ instance Arbitrary AllocationLimitExceeded where
 deriving instance Arbitrary TypeError
 #endif
 
+#if MIN_VERSION_base(4,10,0)
+-- deriving instance Arbitrary CompactionFailed
+instance Arbitrary CompactionFailed where
+    arbitrary = genericArbitrary
+#endif
+
 deriving instance Bounded Deadlock
 deriving instance Enum Deadlock
 instance Arbitrary Deadlock where
@@ -143,6 +149,11 @@ deriving instance Generic RecConError
 deriving instance Generic RecSelError
 deriving instance Generic RecUpdError
 deriving instance Generic ErrorCall
+
+-- TODO: Get rid of this in /Since: next/ version
+# if MIN_VERSION_base(4,10,0)
+deriving instance Generic CompactionFailed
+# endif
 #else
 $(Generics.deriveAll0 ''ArrayException)
 $(Generics.deriveAll0 ''AssertionFailed)